// src/lib/api/generated/model/notificationPreferencePageSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationPreferenceSchema } from './notificationPreferenceSchema';

export interface NotificationPreferencePageSchema {
  items: NotificationPreferenceSchema[];
  page: number;
  page_size: number;
  total: number;
}