// src/lib/api/generated/model/notificationPreferenceSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export interface NotificationPreferenceSchema {
  category: string;
  channel: string;
  description: string;
  email_enabled: boolean;
  in_app_enabled: boolean;
  label: string;
  mandatory: boolean;
  notification_type: string;
  overridden: boolean;
  preference_policy: string;
  priority: string;
}